Q65 generator and demo

+++ mod/fsk.py
@@ -5,7 +5,7 @@ import numpy.typing as npt
 def synth_fsk(
         tones: npt.NDArray[np.int64], sample_rate: int,
         samples_per_symbol: int,
-        f0: float, bandwidth: int) -> npt.NDArray[np.float64]:
+        f0: float, bandwidth: float) -> npt.NDArray[np.float64]:
     dt = 1.0 / sample_rate
 
     phase = 0.0
